refactor: :label: Reverting lat, lng types to float

+++ app/domain/models/marker_model.py
@@ -1,14 +1,11 @@
-from decimal import Decimal
-
-
 class MarkerModel:
     def __init__(
         self,
         marker_id: str,
         incident_id: str,
         direction: str,
-        latitude: Decimal,
-        longitude: Decimal,      
+        latitude: float,
+        longitude: float,      
     ) -> None:
         self.marker_id = marker_id
         self.incident_id = incident_id
