+++ app/infrastructure/dto/marker_dto.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from uuid import UUID
 from pydantic import BaseModel
 
@@ -7,5 +6,5 @@ class MarkerDTO(BaseModel):
     marker_id: UUID
     incident_id: UUID
     direction: str
-    latitude: Decimal
-    longitude: Decimal 
+    latitude: float
+    longitude: float
